:truck: [api]: move libs/lib-blockchain-watcher to libs/blockchain-watcher/feature-app #32
