+++ libs/blockchain-watcher/feature-app/README.md
@@ -1,0 +1,7 @@
+# blockchain-watcher-feature-app
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test blockchain-watcher-feature-app` to execute the unit tests via [Jest](https://jestjs.io).
