+++ libs/blockchain-watcher/feature-app/src/lib/Entity/ScanLogEntity.ts
