+++ libs/blockchain-watcher/feature-app/src/lib/service/mailService.ts
