+++ libs/blockchain-watcher/feature-app/src/lib/store/envStore.ts
