+++ libs/blockchain-watcher/feature-app/src/lib/template/.gitignore
