+++ libs/blockchain-watcher/feature-app/src/lib/template/escrowCreated.mjml
