+++ libs/blockchain-watcher/feature-app/src/lib/util/generate.ts
