--- libs/lib-blockchain-watcher/README.md
@@ -1,7 +1,0 @@
-# lib-blockchain-watcher
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test lib-blockchain-watcher` to execute the unit tests via [Jest](https://jestjs.io).
