--- libs/lib-blockchain-watcher/src/index.ts
@@ -1,7 +1,0 @@
-export * from './lib/Entity/ScanLogEntity'
-export * from './lib/service/dbService'
-export * from './lib/service/ethService'
-export * from './lib/service/mailService'
-export * from './lib/store/envStore'
-export * from './lib/util/generate'
-export * from './lib/util/MailgunTransport'
