+++ tsconfig.base.json
@@ -19,12 +19,15 @@
     "baseUrl": ".",
     "paths": {
       "@assets/*": ["libs/web/ui-assets/src/*"],
-      "@solomon/lib-blockchain-watcher": ["libs/lib-blockchain-watcher/src/index.ts"],
       "@solomon/apps/blockchain-watcher": ["apps/blockchain-watcher/src"],
       "@solomon/apps/developer-docs": ["apps/developer-docs/src"],
+      "@solomon/blockchain-watcher/feature-app": [
+        "libs/blockchain-watcher/feature-app/src/index.ts"
+      ],
       "@solomon/blockchain-watcher/feature-templates": [
         "libs/blockchain-watcher/feature-templates/src/index.ts"
       ],
+      "@solomon/lib-blockchain-watcher": ["libs/lib-blockchain-watcher/src/index.ts"],
       "@solomon/shared/util-api-evidence-types": [
         "libs/shared/util-api-evidence-types/src/index.ts"
       ],
